Implement stations sync task

+++ app/src/main/java/com/udacity/study/jam/radiotastic/station/StationListFragment.java
@@ -133,7 +133,7 @@ public class StationListFragment extends Fragment {
     }
 
     public static interface Callback {
-        void onStationSelected(int stationID);
+        void onStationSelected(double stationID);
     }
 
 }
